requirements.txt
langchain
langchain-openai
openai
python-dotenv
flask
flask-cors
pypdf2
python-docx